Project Blue Sky : Improved service tunnel state management
-- Added a simple state machine in the Connectivity Manager to manage the
state of the service tunnel.  This code now takes into account an new Service
Tunnel Mode properties that controllable by the application.  The logic
also looks at the fabric and service provisioning states of the system to
determine whether the service tunnel should be activated.

-- Implemented platform events that signal changes in the fabric and service
provisioning states of the device.  These used by the Connectivity Manager
to drive changes in the service tunnel state.