Project Blue Sky : Various changes to facilitate pairing
-- Introduced minimal implementation of Network Provisioning TestConnectivity request.
This version does not do a positive check of connectivity to the Internet (e.g. via
sending a HTTP request to a server).  But it does wait for Connectivity Manager to
determine that Internet connectivity is possible by the presence of appropriate local
addresses and configuration.

-- Restructured the way the network provisioning delegate tracks the state of async
operations.  This is in preparation for moving the delegate out of the Connectivity
Manager and into the Network Provisioning Server.