Project Blue Sky : Work on service tunnel management
-- Restructured the way the ConnectivityManager manages the tunnel connection
to the service, creating a distinction between the state of having a tunnel
connection and having “service connectivity”.  This better accommodates the
intermediate state during pairing where the device has a tunnel connection
but the connection is subject to routing restrictions by the service.

-- Introduced a new ServiceTunnelStateChange device event that announces
changes in tunnel connectivity, including whether the tunnel is subject to
routing restrictions.

-- Altered the existing ServiceConnectivityChange event to announce changes
in logical connectivity to the service.  In particular, the ConnectivityManager
now raises the ServiceConnectivityChange/Established event whenever a tunnel
is established that is *not* subject to routing restrictions by the service.

-- Added code to ConnectivityManager to detect when the device completes
account pairing.  At this time, if the tunnel to the service is subject to
routing restrictions (imposed because the device was not paired to an account
at the time the tunnel was established) then the ConnectivityManager will
abort the current tunnel and force a new tunnel to be established.