Project Blue Sky : TraitManager and support for service subscriptions
-- Introduced the TraitManager object to the Weave device layer.  The TraitManager
wraps the Weave WDM APIs and provides a simplified interface for publishing traits
and managing WDM subscriptions to other nodes.  The initial implementation focuses
on support for service mutual subscriptions, providing basic features for
registering trait data sources and sinks, as well as automatic subscription
establishment and maintenance.

-- Created a standard implementation of the Weave DeviceIdentityTrait, which is
automatically published when the device boots.

-- Implemented a data sink for the Weave ApplicationKeysTrait.  TraitManager
automatically includes this trait in the list of trait path when forming a
subscription to the Nest service.

-- Added a ServiceSubscriptionStateChange event, which announces the establishment
or termination of the service mutual subscription to other components in the system
(including the application).

-- Modified the demo app so that the ‘S’ indicator now shows the state of the
service subscription.

-- Added Kconfig options to the demo app to enable/configure the “Alive” message
feature, and the service echo feature. The service echo feature is now disabled by
default.