Make TCP Send Idle Callbacks independent of the TCP UserTimeout configuration (WEAV-3118).
When only TCP Send Idle callback is configured, the SendQueue
would be polled at fixed intervals and upon being flushed, the
application would be notified.

When only TCP UserTimeout is configured, the SendQueue would
only be polled opportunistically(during sends and receives)
and when the TCP UserTimeout happens to determine if the
connection needs to be closed.

When both these configurations are enabled, the TCP sendqueue
is polled at a fixed interval up to a max poll count equivalent
to the TCP User timeout and appropriate actions are taken based
on whether all data is flushed or data flow is
progressing/not progressing.